Creación otros Métodos
Se realiza la creación de los métodos Editar y Eliminar.

+++ api/src/main/java/com/example/restservice/Planet.java
@@ -1,7 +1,4 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
+
 package com.example.restservice;
 
 import jakarta.persistence.Entity;
@@ -16,18 +13,16 @@ import jakarta.persistence.Table;
 class Planet {
     
     @Id
-    //@GeneratedValue(strategy=GenerationType.AUTO)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    //private Integer id;
-    private Long id;
+    private Integer id;
 
     private String name;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
